requirements.txt
langchain
langchain-community
langchain-pinecone
langgraph
pinecone-client
neo4j
sentence-transformers
spacy
python-dotenv
fastapi
uvicorn
pydantic-settings
tiktoken
loguru
tenacity
PyPDF2
pdfplumber
pymupdf
networkx
requests
pytest
httpx
numpy<2.0.0
# Monitoring
langsmith
prometheus-client

# Core Framework
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# pydantic==2.5.0
# pydantic-settings==2.1.0
# python-multipart==0.0.6

# # LLM & Orchestration
# langchain==0.1.0
# langgraph==0.0.20
# langchain-openai==0.0.2
# langchain-community==0.0.10
# openai==1.6.1

# # LlamaIndex (optional - for comparison)
# llama-index==0.9.30

# # Vector Databases
# pinecone-client
# # chromadb==0.4.22
# sentence-transformers==2.2.2

# # Knowledge Graph
# neo4j
# py2neo

# # Document Processing
# pypdf2==3.0.1
# pdfplumber==0.10.3
# python-docx==1.1.0
# pymupdf==1.23.8

# # Text Processing
# spacy==3.7.2
# nltk==3.8.1
# tiktoken==0.5.2

# # Search & Ranking
# rank-bm25==0.2.2

# # Utilities
# python-dotenv==1.0.0
# requests==2.31.0
# aiohttp==3.9.1
# tenacity==8.2.3

# # Monitoring & Logging
# prometheus-client==0.19.0
# loguru==0.7.2

# # Testing
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.25.2

# # Frontend
# streamlit==1.29.0
# plotly==5.18.0
# pandas==2.1.4

# # Development
# black==23.12.1
# flake8==6.1.0
# mypy==1.7.1